backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
